% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernoulli_ARL.R
\name{bernoulli_ARL_MC}
\alias{bernoulli_ARL_MC}
\title{Average run length for Bernoulli CUSUM using Markov Chain methodology}
\usage{
bernoulli_ARL_MC(n_grid, R, h)
}
\arguments{
\item{n_grid}{Number of state spaces used to discretize the outcome space (when \code{method = "MC"})
or number of grid points used for trapezoidal integration (when \code{method = "SPRT"}).
Increasing this number improves accuracy, but can also significantly increase computation time.}

\item{R}{Transition probability matrix obtained from \code{calc_MC_trans_matrix}}

\item{h}{Control limit for the Bernoulli CUSUM}
}
\description{
Internal function that discretizes grid and solves
matrix equation involving transition matrix for Markov Chain methodology
}
\keyword{internal}
