% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgr_control_limit.R
\name{cgr_control_limit}
\alias{cgr_control_limit}
\title{Determine control limits for CGR-CUSUM by simulation}
\usage{
cgr_control_limit(time, alpha = 0.05, psi, n_sim = 20, coxphmod, cbaseh,
  inv_cbaseh, baseline_data, interval = c(0, 9e+12), h_precision = 0.01,
  ncores = 1, seed = 1041996, pb = FALSE, chartpb = FALSE)
}
\arguments{
\item{time}{A numeric value over which the type I error \code{alpha} must be restricted.}

\item{alpha}{A proportion between 0 and 1 indicating the required maximal type I error.
Default is 0.05.}

\item{psi}{A numeric value indicating the estimated Poisson arrival rate of subjects
at their respective units. Can be determined using
\code{\link[cgrcusum:parameter_assist]{parameter_assist()}}.}

\item{n_sim}{An integer value indicating the amount of units to generate for the
determination of the control limit. Larger values yield more precise control limits,
but greatly increase computation times. Default is 20.}

\item{coxphmod}{(optional): A cox proportional hazards regression model as
produced by
the function \code{\link[survival:coxph]{coxph()}}. Suggested: \cr
\code{coxph(Surv(survtime, censorid) ~ covariates, data = baseline_data)}. \cr
Alternatively, a list with the following elements:
\describe{
\item{\code{formula}:}{a \code{\link[stats:formula]{formula()}} in the form \code{~ covariates};}
\item{\code{coefficients}:}{a named vector specifying risk adjustment coefficients
for covariates. Names must be the same as in \code{formula} and colnames of \code{data}.}
}}

\item{cbaseh}{(optional): A function that returns the unadjusted cumulative
baseline hazard \eqn{H_0(t)}{H_0(t)}. If \code{cbaseh} is missing but
\code{coxphmod} has been
specified as a survival object, this baseline hazard rate will be determined
using the provided \code{coxphmod}.}

\item{inv_cbaseh}{(optional): A function that returns the unadjusted inverse cumulative
baseline
hazard \eqn{H^{-1}_0(t)}{H_0^-1(t)}. If \code{inv_cbaseh} is missing, it will be
determined from \code{cbaseh} numerically.}

\item{baseline_data}{(optional): A \code{data.frame} used for covariate resampling
with rows representing subjects and at least the
following named columns: \describe{
\item{\code{entrytime}:}{time of entry into study (numeric);}
\item{\code{survtime}:}{time from entry until event (numeric);}
\item{\code{censorid}:}{censoring indicator (0 = right censored, 1 = observed),
(integer).}
} and optionally additional covariates used for risk-adjustment. Can only be specified
in combination with \code{coxphmod}.}

\item{interval}{(optional): Interval in which survival times should be solved for numerically.}

\item{h_precision}{(optional): A numerical value indicating how precisely the control limit
should be determined. By default, control limits will be determined up to 2 significant digits.}

\item{ncores}{(optional): Number of cores to use to parallelize the computation of the
CGR-CUSUM charts. If ncores = 1 (default), no parallelization is done. You
can use \code{\link[parallel:detectCores]{detectCores()}} to check how many
cores are available on your computer.}

\item{seed}{(optional): A numeric seed for survival time generation. Default = my birthday.}

\item{pb}{(optional): A boolean indicating whether a progress bar should
be shown. Default is \code{FALSE}.}

\item{chartpb}{(optional): A boolean indicating whether progress bars should
be displayed for the constructions of the charts. Default is \code{FALSE}.}
}
\value{
A list containing three components:
\itemize{
\item \code{call}: the call used to obtain output;
\item \code{charts}: A list of length \code{n_sim} containing the constructed charts;
\item \code{data}: A \code{data.frame} containing the in-control generated data.
\item \code{h}: Determined value of the control limit.
}
}
\description{
This function can be used to determine control limits for the
CGR-CUSUM (\code{\link[cgrcusum]{cgr_cusum}}) procedure by restricting the type I error \code{alpha} of the
procedure over \code{time}.
}
\details{
This function performs 3 steps to determine a suitable control limit.
\itemize{
\item Step 1: Generates \code{n_sim} in-control units (failure rate as baseline).
If \code{data} is provided, subject covariates are resampled from the data set.
\item Step 2: Determines chart values for all simulated units.
\item Step 3: Determines control limits such that at most a proportion \code{alpha}
of all units cross the control limit.
} The generated data as well as the charts are also returned in the output.
}
\examples{
require(survival)
\dontrun{
require(survival)
exprfit <- as.formula("Surv(survtime, censorid) ~ age + sex + BMI")
tcoxmod <- coxph(exprfit, data= surgerydat)

a <- cgr_control_limit(time = 500, alpha = 0.1, cbaseh = function(t) chaz_exp(t, lambda = 0.02),
inv_cbaseh = function(t) inv_chaz_exp(t, lambda = 0.02), psi = 0.5, n_sim = 10)

b <- cgr_control_limit(time = 500, alpha = 0.1, coxphmod = tcoxmod, psi = 0.5, n_sim = 10,
data = subset(surgerydat, hosp_num == 1))
}
}
\seealso{
\code{\link[cgrcusum]{plot.cgrcusum}}, \code{\link[cgrcusum]{runlength.cgrcusum}}

Other control limit simulation: 
\code{\link{bernoulli_control_limit}()},
\code{\link{bk_control_limit}()}
}
\author{
Daniel Gomon
}
\concept{control limit simulation}
