% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernoulli_ARL.R
\name{bernoulli_ARL}
\alias{bernoulli_ARL}
\title{Average Run Length for Bernoulli CUSUM}
\usage{
bernoulli_ARL(h, t, glmmod, theta, p0, p1, followup, smooth_prob = FALSE)
}
\arguments{
\item{h}{Control limit for the Bernoulli CUSUM}

\item{t}{Number of state spaced used to discretize the outcome space.}

\item{glmmod}{Generalized linear regression model used for risk-adjustment as produced by
the function \code{\link[stats:glm]{glm()}}. Suggested: \cr
\code{glm(as.formula("(survtime <= followup) & (censorid == 1) ~ covariates"), data = data)}. \cr
Alternatively, a list containing the following elements:
\describe{
\item{\code{formula}:}{a \code{\link[stats:formula]{formula()}} in the form \code{~ covariates};}
\item{\code{coefficients}:}{a named vector specifying risk adjustment coefficients
for covariates. Names must be the same as in \code{formula} and colnames of \code{data}.}
}}

\item{theta}{The \eqn{\theta}{\theta} value used to specify the odds ratio
\eqn{e^\theta}{e^\theta} under the alternative hypothesis.
If \eqn{\theta >= 0}{\theta >= 0}, the chart will try to detect an increase
in hazard ratio (upper one-sided). If \eqn{\theta < 0}{\theta < 0},
the chart will look for a decrease in hazard ratio (lower one-sided).
Note that \deqn{p_1 = \frac{p_0 e^\theta}{1-p_0 +p_0 e^\theta}.}{p1 = (p0 * e^\theta)/(1-p0+p0 * e^\theta).}}

\item{p0}{The baseline failure probability at \code{entrytime + followup} for individuals.}

\item{p1}{The alternative hypothesis failure probability at \code{entrytime + followup} for individuals.}

\item{followup}{(optional) The value of the follow-up time used to determine event time. Will be used to
calculate Average Run Length in time units instead of number of observed outcomes.}

\item{smooth_prob}{Should the probability distribution of failure under the null distribution be smoothed?
Useful for small samples. Can only be TRUE when \code{glmmod} is supplied. Default = FALSE.}
}
\value{
A list containing:
\itemize{
\item \code{ARL_0}: A numeric value indicating the average run length in number of outcomes and
time (if followup specified) when starting from state E_0.
\item \code{ARL}: A \code{data.frame} containing the average run length (#outcomes/time)
depending on the state in which the process starts (E_0, E_1, ..., E_{t-1})
\describe{
\item{\code{t_start}:}{State in which the CUSUM process starts;}
\item{\code{#outcomes}:}{ARL from starting state in #outcomes;}
\item{\code{Time}:}{(only if followup specified) ARL in time units;}
}
\item \code{R}: A transition probability \code{matrix} containing the transition
probabilities between states \eqn{E_0, \ldots, E_{t-1}}{E_0, ..., E_{t-1}}.
\eqn{R_{i,j}}{R_{i,j}} is the transition probability from state i to state j.
}
}
\description{
This function allows to estimate the Average Run Length (ARL)
of the risk-adjusted Bernoulli CUSUM (see \code{\link[success:bernoulli_cusum]{bernoulli_cusum()}})
through a Markov Chain Approach (Brook & Evans(1972) & Steiner et al. (2000)).
The function requires the specification of one of the following combinations of parameters
as arguments to the function:
\itemize{
\item \code{glmmod} & \code{theta}
\item \code{p0} & \code{theta}
\item \code{p0} & \code{p1}
} Average run length of lower-sided Bernoulli CUSUM charts can be determined
by specifying \code{theta} < 0.
}
\details{
The average run length of a CUSUM chart \eqn{S_n}{S_n} is given by
\eqn{\mathbb{E}[\tau_n],}{E[\tau_n],} where \eqn{\tau_n}{\tau_n} is defined as:
\deqn{\tau_n = \inf\{n \geq 0: S_n \geq h\}.}{\tau_n = inf(n >= 0: S_n >= h).}
This function determines the average run length using the Markov Chain approach described
in Brook & Evans (1972), using the risk-adjustment correction proposed in
Steiner et al. (2000). The idea is to discretize the domain (0, h) into $t-1$
state spaces \eqn{E_0, \ldots, E_{t-1}}{E_0, ..., E_{t-1}} of width \eqn{w}{w}, such that
\eqn{E_t}{E_t} is an absorbing state.  This is done using the following steps:
\itemize{
\item \eqn{w}{w} is determined using the relationship \eqn{\frac{2h}{2t-1}}{2h/(2t-1)}.
\item Transition probabilities between the states are determined and
'transition matrix' \eqn{R}{R} is constructed.
\item The equation \eqn{(\bm{I}-\bm{R}) \bm{ARL} = \bm{1}}{(I-R) ARL = 1} is
solved to find the ARL starting from each of the states.
}
}
\examples{
#We consider patient outcomes 100 days after their entry into the study.
followup <- 100
#Determine a risk-adjustment model using a generalized linear model.
#Outcome (failure within 100 days) is regressed on the available covariates:
exprfitber <- as.formula("(survtime <= followup) & (censorid == 1)~ age + sex + BMI")
glmmodber <- glm(exprfitber, data = surgerydat, family = binomial(link = "logit"))
ARL <- bernoulli_ARL(h = 2.5, t = 50, glmmod = glmmodber, theta = log(2))


}
\references{
Brook, D., & Evans, D. A. (1972). An Approach to the Probability
Distribution of Cusum Run Length. Biometrika, 59(3), 539–549.
\doi{10.2307/2334805}

Steiner, S. H., Cook, R. J., Farewell, V. T., & Treasure, T. (2000).
Monitoring surgical performance using risk-adjusted cumulative sum charts.
Biostatistics, 1(4), 441–452. \doi{10.1093/biostatistics/1.4.441}
}
\seealso{
\code{\link[success]{bernoulli_cusum}}, \code{\link[success]{bernoulli_control_limit}}
}
\author{
Daniel Gomon
}
\concept{average run length}
