% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgr_generic.R
\name{plot.cgrcusum}
\alias{plot.cgrcusum}
\alias{plot.bkcusum}
\alias{plot.funnelplot}
\alias{plot.bercusum}
\title{Plot a quality control chart}
\usage{
\method{plot}{cgrcusum}(x, h, ...)

\method{plot}{bkcusum}(x, h, ...)

\method{plot}{funnelplot}(x, percentage = TRUE, ...)

\method{plot}{bercusum}(x, h = x$h, ...)
}
\arguments{
\item{x}{Chart to plot}

\item{h}{Control limit to display for \code{cgrcusum}, \code{bkcusum} or \code{bercusum}}

\item{...}{Further plotting parameters}

\item{percentage}{Should output be shown in percentages? Default is \code{TRUE}.}
}
\value{
A plot of the associated chart is displayed in the current graphics device.
}
\description{
Plot a \code{cgrcusum}, \code{bkcusum},
\code{bercusum} or \code{funnelplot} chart.
}
\section{Methods (by class)}{
\itemize{
\item \code{cgrcusum}: Plot a CGR-CUSUM

\item \code{bkcusum}: Plot a BK-CUSUM

\item \code{funnelplot}: Display a funnel plot

\item \code{bercusum}: Plot a Bernoulli CUSUM
}}

\seealso{
\code{\link[cgrcusum]{cgr_cusum}}, \code{\link[cgrcusum]{bk_cusum}}, \code{\link[cgrcusum]{bernoulli_cusum}}, \code{\link[cgrcusum]{funnel_plot}}
}
\author{
Daniel Gomon
}
