% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgr_summary.R
\name{summary.funnelplot}
\alias{summary.funnelplot}
\title{Summarizes (or elaborates on) S3 objects in this package.}
\usage{
\method{summary}{funnelplot}(object, ...)
}
\arguments{
\item{object}{S3 object to summarize}

\item{...}{extra parameters}
}
\value{
A list with:
\itemize{
\item \code{call}: the call used to obtain the input object,
\item \code{'0.xx'}: the detected instances at specified confidence level(s).
}
}
\description{
Prints the (name of the) instances performing worse than expected
in a \code{funnelplot} object at the confidence levels specified in the
\code{funnelplot} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{funnelplot}: summarize instances detected by the
\code{funnelplot} object
}}

\seealso{
\code{\link[cgrcusum:funnel_plot]{funnel_plot}}
}
